+++ src/user/google/google-user.controller.ts
@@ -1,4 +1,6 @@
 import { Controller } from '@nestjs/common';
 
 @Controller('google')
-export class GoogleController {}
+export class GoogleUserController {
+  
+}
