+++ src/user/google/google-user.service.ts
@@ -1,4 +1,4 @@
 import { Injectable } from '@nestjs/common';
 
 @Injectable()
-export class GoogleService {}
+export class GoogleUserService {}
